third_party/WebKit/LayoutTests/platform/win/fast/css/fontfaceset-add-remove-while-loading-expected.txt

This is a testharness.js-based test.
PASS document.fonts.add() called while a FontFace is loading. 
PASS document.fonts.delete() called while a FontFace is loading. 
PASS document.fonts.clear() called while a FontFace is loading. 
Harness: the test ran to completion.
